backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
diffusers
transformers
accelerate
lpips
clip-openai
scikit-image
ftfy
regex
tqdm
tensorflow
tensorflow-hub
numpy<2
h5py
Pillow
gunicorn
peft
